MOD: 회원가입 API 리팩토링
