+++ src/main/java/com/app/univchat/controller/MemberController.java
@@ -1,22 +1,15 @@
 package com.app.univchat.controller;
 
-import com.app.univchat.base.BaseEntity;
-import com.app.univchat.base.BaseException;
 import com.app.univchat.base.BaseResponse;
 import com.app.univchat.base.BaseResponseStatus;
-import com.app.univchat.dto.MemberDTO;
-import com.app.univchat.repository.MemberRepository;
 import com.app.univchat.service.MemberService;
 import com.app.univchat.dto.MemberReq;
 import com.app.univchat.dto.MemberRes;
 import com.app.univchat.service.EmailService;
+import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.io.IOException;
 
 import static com.app.univchat.base.BaseResponseStatus.*;
 
@@ -40,25 +33,20 @@ public class MemberController {
 
 
     // 회원가입
-
+    @ApiOperation(value = "회원가입 API", notes = "이메일 형식을 보내주세요.")
     @PostMapping("/signup")
-    public BaseResponse<String> signup(@RequestBody MemberDTO memberDTO) throws BaseException {
-        try {
-            // 이메일 중복 체크
-            if(memberService.checkEmail(memberDTO.getEmail())) {
-                return BaseResponse.ok(BaseResponseStatus.USER_ALREADY_EXIST_USERNAME);
-            }
-            // 닉네임 중복 체크
-            if(memberService.checkNickname(memberDTO.getNickname())) {
-                return BaseResponse.ok(BaseResponseStatus.USER_EXISTS_NICKNAME_ERROR);
-            }
+    public BaseResponse<String> signup(@RequestBody MemberReq.Signup memberDto){
 
-//            MemberDTO responseDTO=memberService.signup(memberDTO);
-            return BaseResponse.ok(BaseResponseStatus.SUCCESS,memberService.signup(memberDTO));
-        } catch(BaseException e) {
-
-            return BaseResponse.ok(e.getStatus());
+        // 이메일 중복 체크
+        if(memberService.checkEmail(memberDto.getEmail())) {
+            return BaseResponse.ok(BaseResponseStatus.USER_ALREADY_EXIST_USERNAME);
         }
+        // 닉네임 중복 체크
+        if(memberService.checkNickname(memberDto.getNickname())) {
+            return BaseResponse.ok(BaseResponseStatus.USER_EXISTS_NICKNAME_ERROR);
+        }
+
+        return BaseResponse.ok(BaseResponseStatus.SUCCESS,memberService.signup(memberDto));
 
 
     }
