+++ src/main/java/com/app/univchat/service/MemberService.java
@@ -1,10 +1,8 @@
 package com.app.univchat.service;
 
 import com.app.univchat.base.BaseException;
-import com.app.univchat.dto.MemberDTO;
+import com.app.univchat.dto.MemberReq;
 import com.app.univchat.repository.MemberRepository;
-import lombok.AllArgsConstructor;
-import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -13,7 +11,6 @@ import org.springframework.stereotype.Service;
 //@RequiredArgsConstructor
 public class MemberService {
     private MemberRepository memberRepository;
-//    private PasswordEncoder passwordEncoder;
 
     @Autowired
     public MemberService(MemberRepository memberRepository) {
@@ -21,16 +18,16 @@ public class MemberService {
     }
 
 //    @Transactional
-    public String signup(MemberDTO memberDTO) throws BaseException {
+    public String signup(MemberReq.Signup memberDto) throws BaseException {
 
-        String rawPassword=memberDTO.getPassword();
+        String rawPassword=memberDto.getPassword();
         // password 암호화 - BCryptPasswordEncoder 를 객체로 받아옴
         BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
         String encPassword=passwordEncoder.encode(rawPassword);
-        memberDTO.setPassword(encPassword);
+        memberDto.setPassword(encPassword);
 
         // dto를 entitiy로 변환해 저장
-        memberRepository.save(memberDTO.toEntity());
+        memberRepository.save(memberDto.toEntity());
         return null;
     }
